Use memfd_create syscall instead of libc function
The memfd_create() function is available in glib >= 2.27, but Rust
supports glibc >= 2.17. In order to keep compatibility with old
glibc versions, this performs a memfd_create syscall directly. This
syscall is available in the Linux kernel >= 3.17.

There is currently a pull request
https://github.com/nix-rust/nix/pull/2146
in the nix crate that will use the syscall with all the glibc versions,
in order to account for older glibc versions, while calling the libc
function in other libc's such as musl where the function has existed
from the beginning. When this pull request is merged, this change
can be undone and memfd_create() from nix can be called instead.

Signed-off-by: Daniel Estévez <daniel@destevez.net>